requirements.txt
httpx>=0.27.0
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
flask-limiter>=3.5.0
